refactor: Domain인 chessGame과 DTO를 분리
