+++ src/main/java/chess/console/ConsoleController.java
@@ -3,6 +3,8 @@ package chess.console;
 import chess.domain.ChessGame;
 import chess.domain.command.Command;
 import chess.domain.command.Commands;
+import chess.dto.BoardDto;
+import chess.dto.PointDto;
 import chess.view.InputView;
 import chess.view.OutputView;
 
@@ -21,14 +23,14 @@ public class ConsoleController {
 
     private void view(ChessGame chessGame, Command command) {
         if (command.isStart()) {
-            OutputView.printBoard(chessGame.boardDto());
+            OutputView.printBoard(new BoardDto(chessGame.board(), chessGame.turn()));
         }
         if (command.isMove()) {
-            OutputView.printBoard(chessGame.boardDto());
+            OutputView.printBoard(new BoardDto(chessGame.board(), chessGame.turn()));
             confirmKingDead(chessGame);
         }
         if (command.isStatus()) {
-            OutputView.printStatus(chessGame.pointDto());
+            OutputView.printStatus(new PointDto(chessGame.calculatePoint()));
         }
     }
