+++ src/main/java/chess/dto/CommandDto.java
@@ -1,4 +1,4 @@
-package chess.domain.dto;
+package chess.dto;
 
 public class CommandDto {
     private final String data;
