+++ src/main/java/chess/dto/HistoryDto.java
@@ -1,4 +1,4 @@
-package chess.domain.dto;
+package chess.dto;
 
 public class HistoryDto {
