+++ src/main/java/chess/dto/NameDto.java
@@ -1,4 +1,4 @@
-package chess.domain.dto;
+package chess.dto;
 
 public class NameDto {
     private String name;
