+++ src/main/java/chess/dto/ScoreDto.java
@@ -1,4 +1,4 @@
-package chess.domain.dto;
+package chess.dto;
 
 import chess.domain.board.Team;
