+++ src/main/java/chess/dto/ScoresDto.java
@@ -1,4 +1,4 @@
-package chess.domain.dto;
+package chess.dto;
 
 import java.util.List;
 import java.util.stream.Collectors;
