+++ src/main/java/chess/dto/SquareDto.java
@@ -1,4 +1,4 @@
-package chess.domain.dto;
+package chess.dto;
 
 import chess.domain.piece.Piece;
