+++ src/main/java/chess/service/ChessService.java
@@ -3,11 +3,11 @@ package chess.service;
 import chess.domain.ChessGame;
 import chess.domain.board.Board;
 import chess.domain.command.Commands;
-import chess.domain.dto.CommandDto;
-import chess.domain.dto.GameInfoDto;
-import chess.domain.dto.HistoryDto;
-import chess.domain.repository.CommandRepository;
-import chess.domain.repository.HistoryRepository;
+import chess.dto.CommandDto;
+import chess.dto.GameInfoDto;
+import chess.dto.HistoryDto;
+import chess.repository.CommandRepository;
+import chess.repository.HistoryRepository;
 import chess.domain.utils.PieceInitializer;
 import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
